requirement.txt
datasets
sentence-transformers
torch
faiss-cpu
openai
streamlit
ragas
